regmap: irq: fix ack-invert
When acking irqs we need to take into account the ack-invert case.
Without this chips that require 0's to ACK interrupts will never clear
the interrupt.

By using regmap_irq_update_bits to ACK the interrupts we use the masked
status bits so we take care not to affect any other bits then use
ack_invert to determine if we clear or set those bits.

The only user of ack_invert currently appears to be the motorola-cpcap
driver which we find is incorrectly setting ack_invert and thus we
fix that at the same time as otherwise it would break.

Cc: Guo Zeng <Guo.Zeng@csr.com>
Cc: Benjamin Gaignard <benjamin.gaignard@linaro.org>
Cc: Lee Jones <lee.jones@linaro.org>
Cc: Tony Lindgren <tony@atomide.com>
Signed-off-by: Tim Harvey <tharvey@gateworks.com>