drm/exynos: Fix PLL PMS offset for P value bitfield
The downstream driver for i.MX8MM uses an offset of 14 into the
PLLCTRL register for the P value [1]. Using an offset of 13 as the
Exynos driver does doesn't seem to work. We don't have any
documentation to verify this and therefore can't be sure if
there's a difference in hardware or something else is wrong.

https://source.codeaurora.org/external/imx/linux-imx/tree/drivers/gpu/drm/bridge/sec-dsim.c?h=imx_5.4.47_2.2.0#n211

Signed-off-by: Frieder Schrempf <frieder.schrempf@kontron.de>