net: dev: allow ethernet device index to come from aliases
In order for platform network devices to have determinant names allow
the index to come from dt alias.

Currently this is necessary for the fec and eqos network interfaces on
the GW740x to come up in a determinstic manner without relying on
userspace netdev naming methods such as systemd/udev and/or init
scripts.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>