configs: imx8mm_venice_defconfig: enable ZSTD compression for fw blobs
Newer Linux distros such as Ubuntu 24.04 (Noble) are using
zstd-compressed firmware blobs to reduce filesystem size.

Enable support for this for use on those distro's

While at it remove CONFIG_HW_RANDOM as a cleanup as this is now
configured by default and does not need to be in the defconfig.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>