wifi: nl80211: Introduce the channel flag IEEE80211_CHAN_IGNORE for BCF disabled channels
In this changeset a a new flag IEEE80211_CHAN_IGNORE, which gets applied to
BCF-disabled channels on boot, is introduced. When userspace retrieves wiphy
information via nl80211 this flag is translated to
NL80211_FREQUENCY_ATTR_DISABLED. This allows us to side-step the logic in
cfg80211 that considers wider channels unusable (when IEEE80211_CHAN_DISABLED
is applied), while still presenting a channel as being disabled to userspace.

This change will also enforce (in the supplicant & hostap), that a primary
(or secondary in 5G land) channel is not disabled before tuning to it.

Without this, MorseMicro Halow devices using the out of tree morse_driver
will incorrectly disable some channels.

see
https://github.com/MorseMicro/linux/commit/08f607d73e6be3213deb10f1c25a84d21efe5bf1

Signed-off-by: Tim Harvey <tharvey@gateworks.com>